+++ src/Client/Collection/CollectionGenerator.php
@@ -23,9 +23,9 @@
  * @version     0.8.0
  */
 
-namespace UniAlteri\Sellsy\Client\Collection;
+namespace Teknoo\Sellsy\Client\Collection;
 
-use UniAlteri\Sellsy\Client\Client;
+use Teknoo\Sellsy\Client\Client;
 
 /**
  * Class CollectionGenerator