+++ tests/bootstrap.php
@@ -12,4 +12,4 @@ date_default_timezone_set('UTC');
 
 error_reporting(E_ALL | E_STRICT);
 
-require_once(__DIR__.'/../vendor/autoload.php');
+require_once __DIR__.'/../vendor/autoload.php';